% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-helpers.R
\name{acs_cols_label}
\alias{acs_cols_label}
\title{Variant of \code{gt::cols_label()} for ACS}
\usage{
acs_cols_label(
  data,
  value_col = "estimate",
  value_label = "Est.",
  moe_col = "moe",
  moe_label = "MOE",
  perc_prefix = "perc",
  perc_sep = "_",
  perc_value_label = "\% share",
  perc_moe_label = "\% MOE",
  column_title_col = "column_title",
  column_title_label = NULL,
  env = NULL,
  .col_fn = starts_with
)
}
\arguments{
\item{value_col}{Value column name. Defaults to "estimate"}

\item{value_label}{Value column label. Defaults to "Est.".}

\item{moe_col}{Margin of error column name. Defaults to "moe"}

\item{moe_label}{Margin of error column label. Defaults to "MOE".}

\item{perc_prefix}{Prefix string for percent value columns.}

\item{perc_sep}{Separator string between \code{perc_prefix} and the \code{value_col}
and \code{moe_col} strings.}

\item{perc_value_label, perc_moe_label}{Percent value and margin of error
column labels.}

\item{column_title_col, column_title_label}{Column title column name and
label. Defaults to  "column_title" and \code{NULL}.}

\item{env}{\code{[environment()]}\cr Environment to attach.}

\item{.col_fn}{tidyselect function to use with column names. Defaults to
\link[tidyselect:starts_with]{tidyselect::starts_with},}

\item{variable_col}{Variable column name. Defaults to "variable". Dropped
from table by default.}
}
\description{
Variant of \code{gt::cols_label()} for ACS
}
\keyword{internal}
