% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{acs_table_race_iteration}
\alias{acs_table_race_iteration}
\title{Append race iteration codes to a table ID}
\usage{
acs_table_race_iteration(table, codes = NULL)
}
\arguments{
\item{table}{Table ID.}
}
\description{
Append race iteration codes to a table ID
}
\keyword{internal}
