% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{acs_table_variables}
\alias{acs_table_variables}
\title{Convert a table ID into a set of variable ID using a numeric vector for
variables}
\usage{
acs_table_variables(
  table,
  variables = NULL,
  survey = "acs5",
  year = 2021,
  error_call = caller_env()
)
}
\description{
Convert a table ID into a set of variable ID using a numeric vector for
variables
}
\keyword{internal}
