% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{acs_tables_race_iteration}
\alias{acs_tables_race_iteration}
\title{Get race iteration versions of table string}
\usage{
acs_tables_race_iteration(table)
}
\arguments{
\item{table}{Table ID.}
}
\description{
Get race iteration versions of table string
}
