% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_acs.R
\name{check_acs_id}
\alias{check_acs_id}
\title{Check for presence of ACS ID values}
\usage{
check_acs_id(
  data,
  id = NULL,
  id_col = "table_id",
  allow_any = FALSE,
  allow_null = FALSE,
  id_arg = caller_arg(id),
  arg = caller_arg(data),
  call = caller_env()
)
}
\description{
Check for presence of ACS ID values
}
