% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cli_quiet}
\alias{cli_quiet}
\title{From \code{cliExtras::cli_quiet}}
\usage{
cli_quiet(quiet = FALSE, push = FALSE, .frame = rlang::caller_env())
}
\arguments{
\item{quiet}{If \code{FALSE} (default), leave \code{cli.default_handler} option
unchanged. If \code{TRUE}, set \code{cli.default_handler} to \link{suppressMessages}
temporarily with \code{\link[rlang:local_options]{rlang::local_options()}}}
}
\description{
From \code{cliExtras::cli_quiet}
}
\keyword{internal}
