% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{cols_merge_uncert_ext}
\alias{cols_merge_uncert_ext}
\title{A wrapper for \code{\link[gt:cols_merge_uncert]{gt::cols_merge_uncert()}} with extra features}
\usage{
cols_merge_uncert_ext(
  gt_object,
  columns = NULL,
  col_val = NULL,
  col_uncert = NULL,
  prefix = "",
  postfix = "",
  sep = "",
  ...
)
}
\arguments{
\item{columns}{description}

\item{col_val}{A single column name that contains the base values. This is
the column where values will be mutated.}

\item{col_uncert}{Either one or two column names that contain the uncertainty
values. The most common case involves supplying a single column with
uncertainties; these values will be combined with those in \code{col_val}. Less
commonly, lower and upper uncertainty bounds may be different. For that
case two columns (representing lower and upper uncertainty values away from
\code{col_val}, respectively) should be provided. Since we often don't want the
uncertainty value columns in the output table, we can automatically hide
any \code{col_uncert} columns through the \code{autohide} option.}

\item{prefix, postfix}{Optional strings to insert before and/or after both
col_val and col_uncert. Use a length 2 string \code{c("", "uncert_prefix")} if
you want to apply a prefix to only one or the other column specification.}

\item{sep}{The separator text that contains the uncertainty mark for a single
uncertainty value. The default value of \code{" +/- "} indicates that an
appropriate plus/minus mark will be used depending on the output context.
Should you want this special symbol to be taken literally, it can be
supplied within the \code{\link[=I]{I()}} function.}

\item{...}{One or more character vectors.

\code{NULL}s are removed; scalar inputs (vectors of length 1) are recycled to
the common length of vector inputs.

Like most other R functions, missing values are "infectious": whenever
a missing value is combined with another string the result will always
be missing. Use \code{\link[dplyr:coalesce]{dplyr::coalesce()}} or \code{\link[stringr:str_replace_na]{str_replace_na()}} to convert to
the desired value.}
}
\description{
Wrap \code{\link[gt:cols_merge_uncert]{gt::cols_merge_uncert()}} but allow col_val and col_uncert to be set
based on a length 2 cols parameter and optionally apply a prefix or postfix
value.
}
\keyword{internal}
