% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-helpers.R
\name{.gt_ext}
\alias{.gt_ext}
\title{Create a table with \code{gt::gt()}}
\usage{
.gt_ext(
  data,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  ...,
  drop_geometry = TRUE,
  hide_na_cols = TRUE
)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{rownames_to_stub}{\emph{Use data frame row labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option to take rownames from the input \code{data} table (should they be
available) as row labels in the display table stub.}

\item{row_group_as_column}{\emph{Mode for displaying row group labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option that alters the display of row group labels. By default this is
\code{FALSE} and row group labels will appear in dedicated rows above their
respective groups of rows. If \code{TRUE} row group labels will occupy a
secondary column in the table stub.}

\item{drop_geometry}{If \code{TRUE} (default) and data is an sf object, drop
geometry before turning the data frame into a table.}
}
\description{
Create a table with \code{gt::gt()}
}
\keyword{internal}
