% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_recode_with_list}
\alias{fct_recode_with_list}
\title{Helper for recoding based on a named list}
\usage{
fct_recode_with_list(x, list = NULL, in_order = TRUE, ordered = NA)
}
\description{
Helper for recoding based on a named list
}
\keyword{internal}
