% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_acs.R
\name{filter_acs}
\alias{filter_acs}
\title{Filter American Community Survey data by table, variables, or other attributes}
\usage{
filter_acs(
  data,
  ...,
  table = NULL,
  column = NULL,
  vars = NULL,
  drop_vars = NULL
)
}
\arguments{
\item{data}{A data frame with a "table_id", "variable", and "column_title"
columns.}

\item{...}{Parameters passed to \code{\link[dplyr:filter]{dplyr::filter()}}}

\item{table, column}{Table ID and column title values to return.}

\item{vars, drop_vars}{Variables to keep and drop.}
}
\description{
A data frame of American Community Survey data enriched with table and column
metadata using the \code{\link[=label_acs_metadata]{label_acs_metadata()}} function.
}
\seealso{
\code{\link[dplyr:filter]{dplyr::filter()}}
}
\keyword{internal}
