% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_fmt.R
\name{fmt_acs_county_col}
\alias{fmt_acs_county_col}
\alias{fmt_acs_county}
\title{Assorted formatting functions for ACS tables}
\usage{
fmt_acs_county(
  data,
  state,
  pattern = ", {state}",
  name_col = "NAME",
  columns = dplyr::all_of(name_col),
  ...
)
}
\description{
Assorted formatting functions for ACS tables
}
