% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{fmt_acs_estimate}
\alias{fmt_acs_estimate}
\title{Format estimate and MOE columns in a gt table}
\usage{
fmt_acs_estimate(
  gt_object,
  col_est = "estimate",
  col_moe = "moe",
  columns = NULL,
  col_labels = "Est.",
  spanner = NULL,
  decimals = 0,
  use_seps = TRUE
)
}
\description{
Format estimate and MOE columns in a gt table
}
\keyword{internal}
