% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_acs.R
\name{fmt_acs_minutes}
\alias{fmt_acs_minutes}
\title{Format ACS columns using different units}
\usage{
fmt_acs_minutes(data, column_title = "column_title")
}
\description{
Format ACS columns using different units
}
\keyword{internal}
