% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{fmt_acs_percent}
\alias{fmt_acs_percent}
\title{Format \% estimate and \% MOE columns in a gt table}
\usage{
fmt_acs_percent(
  gt_object,
  col_est = "perc_estimate",
  col_moe = "perc_moe",
  columns = NULL,
  col_labels = "\% share",
  spanner = NULL,
  decimals = 0,
  use_seps = TRUE,
  ...
)
}
\description{
Format \% estimate and \% MOE columns in a gt table
}
\keyword{internal}
