% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{fmt_acs_values}
\alias{fmt_acs_values}
\title{Apply formatting to value and percent value columns using \code{gt::fmt_number()}
and \code{gt::fmt_percent()}}
\usage{
fmt_acs_values(
  data,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  decimals = 0,
  use_seps = TRUE,
  accounting = FALSE,
  scale_by = 1,
  merge_moe = TRUE,
  currency_value = FALSE,
  ...,
  .cols_fn = starts_with
)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[gt:gt]{gt()}} function.}

\item{value_col}{Column name for estimate value column. Defaults to
"estimate".}

\item{moe_col}{Column name for margin of error column. Defaults to
"moe".}

\item{perc_prefix}{Prefix string for percent value columns.}

\item{perc_sep}{Separator string between \code{perc_prefix} and the \code{value_col}
and \code{moe_col} strings.}

\item{decimals}{\emph{Number of decimal places}

\code{scalar<numeric|integer>(val>=0)} // \emph{default:} \code{2}

This corresponds to the exact number of decimal places to use. A value
such as \code{2.34} can, for example, be formatted with \code{0} decimal places and
it would result in \code{"2"}. With \code{4} decimal places, the formatted value
becomes \code{"2.3400"}.}

\item{use_seps}{\emph{Use digit group separators}

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

An option to use digit group separators. The type of digit group separator
is set by \code{sep_mark} and overridden if a locale ID is provided to \code{locale}.
This setting is \code{TRUE} by default.}

\item{accounting}{\emph{Use accounting style}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option to use accounting style for values. Normally, negative values
will be shown with a minus sign but using accounting style will instead put
any negative values in parentheses.}

\item{scale_by}{\emph{Scale values by a fixed multiplier}

\verb{scalar<numeric|integer>} // \emph{default:} \code{1}

All numeric values will be multiplied by the \code{scale_by} value before
undergoing formatting. Since the \code{default} value is \code{1}, no values will be
changed unless a different multiplier value is supplied. This value will be
ignored if using any of the \code{suffixing} options (i.e., where \code{suffixing} is
not set to \code{FALSE}).}

\item{merge_moe}{If \code{TRUE}, use \code{\link[gt:cols_merge_uncert]{gt::cols_merge_uncert()}} to merge the
value_col and moe_col and the percent value and margin of error columns.}

\item{currency_value}{If \code{TRUE}, use \code{\link[gt:fmt_currency]{gt::fmt_currency()}} to format value
columns instead of \code{\link[gt:fmt_number]{gt::fmt_number()}}.}

\item{...}{Additional parameters passed to \code{\link[gt:fmt_currency]{gt::fmt_currency()}} if
\code{currency_value = TRUE}}

\item{.cols_fn}{tidyselect function used to select columns to format and
merge. Defaults to \link[tidyselect:starts_with]{tidyselect::starts_with}}
}
\description{
\code{\link[=fmt_acs_values]{fmt_acs_values()}} is a wrapper for \code{\link[gt:fmt_number]{gt::fmt_number()}}, \code{\link[gt:fmt_percent]{gt::fmt_percent()}},
and \code{\link[gt:cols_merge_uncert]{gt::cols_merge_uncert()}}. If \code{value_col} starts with the string supplied
to perc_prefix, the value column is formatted as a percent value.
}
\keyword{internal}
