% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ts.R
\name{get_acs_ts}
\alias{get_acs_ts}
\title{Get a time series of ACS data}
\usage{
get_acs_ts(
  geography,
  variables = NULL,
  table = NULL,
  cache_table = TRUE,
  year = 2021,
  state = NULL,
  county = NULL,
  survey = "acs5",
  ...
)
}
\arguments{
\item{geography}{The geography of your data.}

\item{variables}{Character string or vector of character strings of variable
IDs. tidycensus automatically returns the estimate and the margin of error
associated with the variable.}

\item{table}{A character vector of tables.}

\item{cache_table}{Whether or not to cache table names for faster future
access. Defaults to FALSE; if TRUE, only needs to be called once per
dataset.  If variables dataset is already cached via the
\code{load_variables} function, this can be bypassed.}

\item{year}{If length 1, the function uses \code{\link[=acs_survey_ts]{acs_survey_ts()}} to get data for
all comparable survey years back to the start of the ACS. If length is
greater than 1, return the selected years even if those years may not be
valid to compare.}

\item{state}{An optional vector of states for which you are requesting data.
State names, postal codes, and FIPS codes are accepted. Defaults to NULL.}

\item{county}{The county for which you are requesting data. County names and
FIPS codes are accepted. Must be combined with a value supplied to `state`.
Defaults to NULL.}

\item{survey}{The ACS contains one-year, three-year, and five-year surveys
expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}

\item{...}{Other keyword arguments}
}
\description{
Get a time series of ACS data
}
