% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decennial_ts.R
\name{get_decennial_ts}
\alias{get_decennial_ts}
\title{Get multiple years of decennial US Census data for time series analysis}
\usage{
get_decennial_ts(
  geography,
  variables = NULL,
  table = NULL,
  cache_table = TRUE,
  year = 2020,
  sumfile = NULL,
  state = NULL,
  county = NULL,
  geometry = FALSE,
  summary_var = NULL,
  label = TRUE,
  ...
)
}
\arguments{
\item{geography}{The geography of your data.}

\item{variables}{If any year value is 2020, variables must be the same length
as year with each value corresponding to one of the years requested. This
is a temporary requirement to address the mismatch between the available
data for 2000 and 2010 relative to 2020.
Default: \code{NULL}}

\item{table}{The Census table for which you would like to request all variables. Uses
lookup tables to identify the variables; performs faster when variable
table already exists through \code{load_variables(cache = TRUE)}.
Only one table may be requested per call.}

\item{cache_table}{Whether or not to cache table names for faster future access.
Defaults to FALSE; if TRUE, only needs to be called once per
dataset.  If variables dataset is already cached via the
\code{load_variables} function, this can be bypassed.}

\item{year}{If year is length 1, it is treated as the max year and decennial
Census years back to 2000, are added to the vector of requested years.
Default: 2020}

\item{sumfile}{The Census summary file; if \code{NULL}, defaults to \code{"pl"} when the year is 2020 and \code{"sf1"} for 2000 and 2010.  Not all summary files are available for each decennial Census year.  Make sure you are using the correct summary file for your requested variables, as variable IDs may be repeated across summary files and represent different topics.}

\item{state}{The state for which you are requesting data. State
names, postal codes, and FIPS codes are accepted.
Defaults to NULL.}

\item{county}{The county for which you are requesting data. County names and
FIPS codes are accepted. Must be combined with a value supplied
to `state`.  Defaults to NULL.}

\item{geometry}{if FALSE (the default), return a regular tibble of ACS data.
if TRUE, uses the tigris package to return an sf tibble
with simple feature geometry in the `geometry` column.}

\item{summary_var}{Character string of a "summary variable" from the decennial Census
to be included in your output. Usually a variable (e.g. total population)
that you'll want to use as a denominator or comparison.}

\item{label}{If \code{TRUE} (default), use \code{\link[=label_decennial_data]{label_decennial_data()}} to add
formatted label columns to the decennial Census data frame.}

\item{...}{
  Arguments passed on to \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial}}
  \describe{
    \item{\code{output}}{One of "tidy" (the default) in which each row represents an
enumeration unit-variable combination, or "wide" in which each
row represents an enumeration unit and the variables are in the
columns.}
    \item{\code{keep_geo_vars}}{if TRUE, keeps all the variables from the Census
shapefile obtained by tigris.  Defaults to FALSE.}
    \item{\code{shift_geo}}{(deprecated) if TRUE, returns geometry with Alaska and Hawaii
shifted for thematic mapping of the entire US.
Geometry was originally obtained from the albersusa R package.  As of May 2021,
we recommend using \code{tigris::shift_geometry()} instead.}
    \item{\code{pop_group}}{The population group code for which you'd like to request data.  Applies to summary files for which population group breakdowns are available like the Detailed DHC-A file.}
    \item{\code{pop_group_label}}{If \code{TRUE}, return a \code{"pop_group_label"} column that contains the label for the population group.  Defaults to \code{FALSE}.}
    \item{\code{key}}{Your Census API key.
Obtain one at \url{https://api.census.gov/data/key_signup.html}}
    \item{\code{show_call}}{if TRUE, display call made to Census API. This can be very useful
in debugging and determining if error messages returned are
due to tidycensus or the Census API. Copy to the API call into
a browser and see what is returned by the API directly. Defaults to FALSE.}
  }}
}
\value{
A data frame with decennial Census data.
}
\description{
\code{\link[=get_decennial_ts]{get_decennial_ts()}} is a wrapper for \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}} to handle
time series data.
}
\examples{
\dontrun{
if (interactive()) {
  md_counties <- get_decennial_ts(
    geography = "county",
    variables = c("P001001", "P001001", "P1_001N"),
    year = 2020,
    county = "Baltimore city",
    state = "MD",
    geometry = FALSE
  )
}
}
}
\seealso{
\code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}
}
