% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{get_geography_params}
\alias{get_geography_params}
\title{Get geography parameters}
\usage{
get_geography_params(
  geography,
  year = 2021,
  state = NULL,
  county = NULL,
  allow_decennial = FALSE,
  call = caller_env()
)
}
\arguments{
\item{geography}{The geography of your data.}

\item{year}{Survey year.}

\item{state, county}{State and county. Defaults to \code{NULL}.}

\item{allow_decennial}{If \code{TRUE}, allow geography values of "block" or
"voting district" that are only supported by \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.
If \code{FALSE} (default), error on those geographies that are not supported by
\code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Get a named list of parameters including geography and, optionally, state and
county, which are dropped or checked depending on whether the value is
required to download data at the specified geography using the Census API.
}
\keyword{internal}
