% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACS-package.R
\docType{package}
\name{getACS-package}
\alias{getACS}
\alias{getACS-package}
\title{getACS: Help Wrangling Data from tidycensus}
\description{
A package with helper functions for working with Census data downloaded with the {tidycensus} package.
}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
