% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACSdata.R
\name{getACS1vars}
\alias{getACS1vars}
\title{Load list of ACS 1-year variables}
\usage{
getACS1vars(year)
}
\arguments{
\item{year}{ACS 1-year year}
}
\value{
A data.table with list of vars
}
\description{
Load list of ACS 1-year variables
}
