% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACSdata.R
\name{getACS5vars}
\alias{getACS5vars}
\title{Load list of ACS 5-year variables}
\usage{
getACS5vars(year)
}
\arguments{
\item{year}{ACS 5-year year}
}
\value{
A data.table with list of vars
}
\description{
Load list of ACS 5-year variables
}
