% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACSdata.R
\name{getACSdata}
\alias{getACSdata}
\title{Get ACS Data using API}
\usage{
getACSdata(vars, endpoint, apikey, geounit, ...)
}
\arguments{
\item{vars}{List of ACS variable names}

\item{endpoint}{ACS API endpoint}

\item{apikey}{API Key}

\item{geounit}{API Geography unit}
}
\value{
A data.table with data returned by API
}
\description{
Creates API call for ACS data.
}
