% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACSdata.R
\name{getACSgroup}
\alias{getACSgroup}
\title{Return character vector of variable names given ACS group}
\usage{
getACSgroup(group, vars)
}
\arguments{
\item{group}{ACS group name}

\item{vars}{List of ACS vars from getACS#vars()}
}
\value{
a character vector of variable names
}
\description{
Return character vector of variable names given ACS group
}
