% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_acs.R
\name{gt_acs}
\alias{gt_acs}
\title{Create a gt table with formatted ACS estimate and percent estimate columns}
\usage{
gt_acs(
  data,
  rownames_to_stub = FALSE,
  row_group_as_column = FALSE,
  ...,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  column_title_col = "column_title",
  name_col = "NAME",
  perc_value_label = "\% share",
  value_label = "Est.",
  column_title_label = NULL,
  name_label = NULL,
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  decimals = 0,
  source_note = NULL,
  append_note = FALSE,
  drop_geometry = TRUE,
  hide_na_cols = TRUE,
  currency_value = FALSE,
  survey = "acs5",
  year = 2021,
  table = NULL,
  prefix = "Source: ",
  end = ".",
  est_cols = NULL,
  perc_cols = NULL
)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{rownames_to_stub}{\emph{Use data frame row labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option to take rownames from the input \code{data} table (should they be
available) as row labels in the display table stub.}

\item{row_group_as_column}{\emph{Mode for displaying row group labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option that alters the display of row group labels. By default this is
\code{FALSE} and row group labels will appear in dedicated rows above their
respective groups of rows. If \code{TRUE} row group labels will occupy a
secondary column in the table stub.}

\item{...}{Additional parameters passed to \code{\link[gt:fmt_number]{gt::fmt_number()}} by
\code{\link[=fmt_acs_estimate]{fmt_acs_estimate()}} or to \code{\link[gt:fmt_percent]{gt::fmt_percent()}} by \code{\link[=fmt_acs_percent]{fmt_acs_percent()}}.}

\item{value_col}{Value column name. Defaults to "estimate"}

\item{moe_col}{Margin of error column name. Defaults to "moe"}

\item{perc_prefix}{Prefix string for percent value columns.}

\item{perc_sep}{Separator string between \code{perc_prefix} and the \code{value_col}
and \code{moe_col} strings.}

\item{column_title_col, column_title_label}{Column title and label. If
\code{column_title_label} is a string, \code{column_title_col} is required.
\code{column_title_label} can also be a named vector in the format of \code{c("label" = "column")}. \code{column_title_col} defaults to "column_title". If
\code{column_title_label} is "from_table", the label is set based on the
simple_table_title column in the table metadata.}

\item{name_col, name_label}{Place name column and label. \code{name_label}
can be a string or a named vector (similar to \code{column_title_label}).
\code{name_col} defaults to "NAME"}

\item{perc_value_label}{Percent value column label.}

\item{value_label}{Value column label. Defaults to "Est.".}

\item{est_spanner, perc_spanner}{Spanner labels for estimate and percent
estimate columns.}

\item{combined_spanner}{If not \code{NULL}, combined_spanner is passed to label
parameter of \code{\link[gt:tab_spanner]{gt::tab_spanner()}} using the value columns and percent
columns as the columns parameter.}

\item{decimals}{\emph{Number of decimal places}

\code{scalar<numeric|integer>(val>=0)} // \emph{default:} \code{2}

This corresponds to the exact number of decimal places to use. A value
such as \code{2.34} can, for example, be formatted with \code{0} decimal places and
it would result in \code{"2"}. With \code{4} decimal places, the formatted value
becomes \code{"2.3400"}. The trailing zeros can be removed with
\code{drop_trailing_zeros = TRUE}. If you always need \code{decimals = 0}, the
\code{\link[gt:fmt_integer]{fmt_integer()}} function should be considered.}

\item{source_note}{\emph{Source note text}

\verb{scalar<character>} // \strong{required}

Text to be used in the source note. We can optionally use the \code{\link[gt:md]{md()}} and
\code{\link[gt:html]{html()}} functions to style the text as Markdown or to retain HTML elements
in the text.}

\item{append_note}{If \code{TRUE}, add source_note to the end of the generated ACS
data label. If \code{FALSE}, any supplied source_note will be used instead of an
ACS label.}

\item{drop_geometry}{If \code{TRUE} (default) and data is an sf object, drop
geometry before turning the data frame into a table.}

\item{hide_na_cols}{If \code{TRUE} (default), hide columns where all values are
\code{NA}.}

\item{currency_value}{If \code{TRUE}, use \code{\link[gt:fmt_currency]{gt::fmt_currency()}} to format value
columns instead of \code{\link[gt:fmt_number]{gt::fmt_number()}}.}

\item{survey}{ACS survey, "acs5", "acs3", or "acs1".}

\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}

\item{table}{One or more table IDs to include in label or source note.}

\item{prefix}{Text to insert before ACS survey label.}

\item{end}{A character string appended to the end of the full label. Defaults
to ".".}

\item{est_cols, perc_cols}{Deprecated. Estimate and percent estimate columns.}
}
\description{
Create or format a gt table with an estimate and margin of error and
(optionally) percent estimate and margin of error value. Use in combination
with the \code{\link[=select_acs]{select_acs()}} helper function to prep data before creating a
table.
}
\examples{
\dontrun{
if (interactive()) {
  data <- get_acs_tables(
    geography = "county",
    county = "Baltimore city",
    state = "MD",
    table = "B08134"
  )

  tbl_data <- filter_acs(data, indent == 1, line_number <= 10)
  tbl_data <- select_acs(tbl_data)

  gt_acs(
    tbl_data,
    column_title_label = "Commute time",
    table = "B08134"
  )
}
}
}
\seealso{
Other gt table: 
\code{\link{fmt_acs_estimate}()},
\code{\link{gt_acs_compare}()},
\code{\link{tab_acs_source_note}()}
}
\concept{gt table}
\keyword{gt}
