% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_acs_compare.R
\name{gt_acs_compare}
\alias{gt_acs_compare}
\title{Create a gt table with values compared by name, geography, or other grouping variable}
\usage{
gt_acs_compare(
  data,
  .by = name_col,
  name_col = "NAME",
  value_cols = c("estimate", "moe"),
  perc_value_cols = c("perc_estimate", "perc_moe"),
  column_title_col = "column_title",
  variable_col = "variable",
  drop_cols = c(name_col, variable_col, "GEOID"),
  value_col_label = "Est.",
  perc_value_col_label = "\% share",
  decimals = 0,
  return_gt = TRUE,
  ...
)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{.by}{Name of grouping column. Defaults to \code{name_col}}

\item{name_col}{Name column. Defaults to "NAME". Dropped from table by
default.}

\item{variable_col}{Variable column name. Defaults to "variable". Dropped
from table by default.}

\item{drop_cols}{Columns to drop before creating gt table. Defaults to
\code{c(name_col, variable_col, "GEOID")}}

\item{value_col_label, perc_value_col_label}{Labels for value/estimate column
and percent value/estimate column. Not yet implemented as an option.}

\item{return_gt}{If \code{TRUE} (default), return a formatted gt table. If
\code{FALSE}, return the data frame for a comparison table.}

\item{...}{
  Arguments passed on to \code{\link[=gt_acs]{gt_acs}}
  \describe{
    \item{\code{combined_spanner}}{If not \code{NULL}, combined_spanner is passed to label
parameter of \code{\link[gt:tab_spanner]{gt::tab_spanner()}} using both est_cols and perc_cols as the
columns parameter.}
    \item{\code{column_title_col,column_title_label}}{Column title and label. If
column_title_label is a string, column_title_col is required.
column_title_label can also be a named vector in the format of \code{c("label" = "column")}. column_title_col defaults to "column_title"}
    \item{\code{name_col,name_col_label}}{Place name column and label. name_col_label
can be a string or a named vector (similar to column_title_label). name_col
defaults to "NAME"}
    \item{\code{hide_na_cols}}{If \code{TRUE} (default), hide columns where all values are
\code{NA}.}
    \item{\code{drop_geometry}}{If \code{TRUE} (default) and data is an sf object, drop
geometry before turning the data frame into a table.}
    \item{\code{rownames_to_stub}}{\emph{Use data frame row labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option to take rownames from the input \code{data} table (should they be
available) as row labels in the display table stub.}
    \item{\code{row_group_as_column}}{\emph{Mode for displaying row group labels in the stub}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option that alters the display of row group labels. By default this is
\code{FALSE} and row group labels will appear in dedicated rows above their
respective groups of rows. If \code{TRUE} row group labels will occupy a
secondary column in the table stub.}
    \item{\code{append_note}}{If \code{TRUE}, add source_note to the end of the generated ACS
data label. If \code{FALSE}, any supplied source_note will be used instead of an
ACS label.}
    \item{\code{source_note}}{\emph{Source note text}

\verb{scalar<character>} // \strong{required}

Text to be used in the source note. We can optionally use the \code{\link[gt:md]{md()}} and
\code{\link[gt:html]{html()}} functions to style the text as Markdown or to retain HTML elements
in the text.}
    \item{\code{survey}}{ACS survey, "acs5", "acs3", or "acs1".}
    \item{\code{year}}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}
    \item{\code{prefix}}{Text to insert before ACS survey label.}
    \item{\code{table}}{One or more table IDs to include in label or source note.}
    \item{\code{end}}{A character string appended to the end of the full label. Defaults
to ".".}
  }}
}
\description{
\code{\link[=gt_acs_compare]{gt_acs_compare()}} is a variant of \code{\link[=gt_acs]{gt_acs()}} designed to support the
comparison of multiple named areas or multiple geographies side-by-side in a
combined gt table. This function \emph{only} works if the arrangement of rows is
the same for each group and all groups have the same number of rows.
}
