% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jam_values}
\alias{jam_values}
\title{ACS Jam Values for Medians}
\format{
A data frame with 20 rows and 6 variables:
\describe{
\item{\code{value}}{Estimate value}
\item{\code{meaning}}{Meaning of estimate value}
\item{\code{use}}{Subjects/tables where jam value is used}
\item{\code{type}}{Type (minimum or maximum jam value)}
\item{\code{units}}{Units. Note year is for a specific year, years is for duration.}
\item{\code{year}}{Year applicable}
}
}
\usage{
jam_values
}
\description{
Reference table of ACS "jam values" for medians from "Table 5.2. Jam Values
for Medians," \emph{Understanding and Using American Community Survey Data: What
All Data Users Need to Know} (2020). \code{type} and \code{units} values are added.
\code{year} is included to account for the possibility of alternate jam values for
earlier or later years but annual variation in values has not been checked.
}
\details{
\url{https://docs.google.com/spreadsheets/d/1YX3NBDkkoDXHs88KDfPS_QoS9-1j_C_q8UAyjPznfzA/edit?usp=sharing}
}
\keyword{datasets}
