% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_acs.R
\name{join_acs_geography_ratio}
\alias{join_acs_geography_ratio}
\title{Join ACS data from a single reference geography by variable to calculate a
ratio value based on the reference geography data}
\usage{
join_acs_geography_ratio(
  data,
  variable_col = "variable",
  value_col = "estimate",
  moe_col = "moe",
  geography = "county",
  na_matches = "never",
  digits = 2
)
}
\arguments{
\item{data}{A data frame with column names matching the supplied parameters.}

\item{value_col, moe_col}{Estimate and margin of error column names,
Default: 'estimate' and 'moe'}

\item{geography}{Value in geography column to use as comparison values,
Default: 'county'}

\item{na_matches}{Should two \code{NA} or two \code{NaN} values match?
\itemize{
\item \code{"na"}, the default, treats two \code{NA} or two \code{NaN} values as equal, like
\code{\%in\%}, \code{\link[=match]{match()}}, and \code{\link[=merge]{merge()}}.
\item \code{"never"} treats two \code{NA} or two \code{NaN} values as different, and will
never match them together or to any other values. This is similar to joins
for database sources and to \code{base::merge(incomparables = NA)}.
}}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}

\item{column_col}{Variable column name to join as join variable, Default:
'variable'}
}
\value{
A data frame with new estimate and moe columns prefixed with
"ratio_".
}
\description{
\code{\link[=join_acs_prop]{join_acs_prop()}} uses data from \code{\link[=get_acs_geographies]{get_acs_geographies()}} to support the
calculation of proportions join parent column titles to a data frame of ACS
data.
}
\seealso{
\code{\link[tidycensus:moe_ratio]{tidycensus::moe_ratio()}}
}
