% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decennial_ts.R
\name{label_decennial_data}
\alias{label_decennial_data}
\title{Label Decennial Census data}
\usage{
label_decennial_data(data, year = 2020, dataset = NULL, ...)
}
\arguments{
\item{year}{The year for which you are requesting variables. Either the year
or endyear of the decennial Census or ACS sample. 5-year ACS data is
available from 2009 through 2020. 1-year ACS data is available from 2005
through 2021, with the exception of 2020.}

\item{dataset}{The dataset name as used on the Census website.  See the Details in this documentation for a full list of dataset names.}

\item{...}{
  Arguments passed on to \code{\link[=load_decennial_vars]{load_decennial_vars}}
  \describe{
    \item{\code{cache}}{Whether you would like to cache the dataset for future access,
or load the dataset from an existing cache. Defaults to FALSE.}
  }}
}
\description{
\code{\link[=label_decennial_data]{label_decennial_data()}} left joins the data frame of variables from
\code{\link[=load_decennial_vars]{load_decennial_vars()}} to the input data.
}
\keyword{internal}
