% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_decennial_ts.R
\name{load_decennial_vars}
\alias{load_decennial_vars}
\title{Load decennial variables}
\usage{
load_decennial_vars(year = 2020, dataset = NULL, cache = TRUE, quiet = FALSE)
}
\arguments{
\item{year}{The year for which you are requesting variables. Either the year
or endyear of the decennial Census or ACS sample. 5-year ACS data is
available from 2009 through 2020. 1-year ACS data is available from 2005
through 2021, with the exception of 2020.}

\item{dataset}{The dataset name as used on the Census website.  See the Details in this documentation for a full list of dataset names.}

\item{cache}{Whether you would like to cache the dataset for future access,
or load the dataset from an existing cache. Defaults to FALSE.}
}
\description{
\code{\link[=load_decennial_vars]{load_decennial_vars()}} downloads metadata for the specified year and dataset
using \code{\link[tidycensus:load_variables]{tidycensus::load_variables()}} then creates new \code{table_id}, \code{indent},
\code{line_number}, and \code{column_title} columns similar to those provided in the
prepared Census Reporter metadata used by \code{\link[=label_acs_metadata]{label_acs_metadata()}}.
}
\keyword{internal}
