% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{make_area_xwalk}
\alias{make_area_xwalk}
\alias{use_area_xwalk}
\title{Make and use crosswalk data based on U.S. Census block-level weights for U.S.
Census tracts and non-Census geographic areas}
\usage{
make_area_xwalk(
  area,
  block_xwalk = NULL,
  state = NULL,
  county = NULL,
  year = 2020,
  name_col = "NAME",
  weight_col = "HOUSING20",
  geoid_col = "GEOID",
  tract_col = "TRACTCE20",
  by = c(TRACTCE20 = "TRACTCE"),
  suffix = c("_block", "_tract"),
  placement = c("largest", "surface", "centroid"),
  digits = 2,
  extensive = TRUE,
  coverage = TRUE,
  erase = FALSE,
  area_threshold = 0.75,
  keep_geometry = FALSE,
  crs = NULL,
  make_valid = TRUE,
  ...
)

use_area_xwalk(
  data,
  area_xwalk,
  geography = "area",
  name_col = "NAME",
  geoid_col = "GEOID",
  suffix = c("_area", ""),
  weight_col = "perc_HOUSING20",
  variable_col = "variable",
  value_col = "estimate",
  moe_col = "moe",
  digits = 0,
  perc = TRUE,
  extensive = TRUE,
  reliability = FALSE,
  moe_level = 90,
  call = caller_env()
)
}
\arguments{
\item{area}{A sf object with an arbitrary geography overlapping with the
block_xwalk. Required. If area only partly overlaps with block_xwalk,
coverage should be set to \code{TRUE} (default).}

\item{block_xwalk}{Block-tract crosswalk sf object. If \code{NULL}, state is
required to create a crosswalk using \code{\link[=make_block_xwalk]{make_block_xwalk()}}}

\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties.
Can also be a county name or vector of names.}

\item{year}{the data year; defaults to 2024}

\item{name_col}{Name column in area.}

\item{weight_col}{Column name or names in input block_xwalk to use for
weighting. Generated weight_col used by \code{\link[=use_area_xwalk]{use_area_xwalk()}} should be the
same as the weight_col for \code{\link[=make_area_xwalk]{make_area_xwalk()}} but include the "perc_"
prefix. Defaults to "HOUSING20" for \code{\link[=make_block_xwalk]{make_block_xwalk()}} and
"perc_HOUSING20" for \code{\link[=use_area_xwalk]{use_area_xwalk()}}.}

\item{geoid_col, tract_col}{GeoID for Census tract and Census tract ID column
in block_xwalk}

\item{by}{Specification of join variables in the format of c("block column
name for tract" = "tract column name"). Passed to \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}.}

\item{suffix}{Suffixes added to the output to disambiguate column names from
the block and tract data. Unused for 2020 data.}

\item{placement}{String with option for joining \code{area} and \code{block_xwalk}:
"largest", "surface", or "centroid". "largest" joins the two using
\code{\link[sf:st_join]{sf::st_join()}} with largest set to \code{TRUE}. "surface" first transforms
block_xwalk using \code{\link[sf:geos_unary]{sf::st_point_on_surface()}} and "centroid" uses
\code{\link[sf:geos_unary]{sf::st_centroid()}}.}

\item{digits}{Digits to use for percent share of weight value.}

\item{extensive}{If \code{TRUE} (default) calculate new estimate values as
weighted sums and re-calculate margin of error with
\code{\link[tidycensus:moe_sum]{tidycensus::moe_sum()}}. If \code{FALSE}, calculate new estimate values as
weighted means (appropriate for ACS median variables) and drop the margin
of error. \code{perc} is also always set to \code{FALSE} if extensive is \code{FALSE}.}

\item{coverage}{If \code{TRUE} (default), it is assumed that area does not cover
the full extent of the block_xwalk and an additional feature is added with
the difference between the unioned area geometry and unioned block_xwalk
geometry. This additional coverage ensures that blocks are accurately
assigned to this alternate geography but it is excluded from the returned
data frame. If \code{coverage} is \code{TRUE} and all features in area overlap with
block_xwalk, the function issues a warning and then resets coverage to
\code{FALSE}. The reverse option is applied if any features from area do not
overlap. \code{coverage} can also be a \code{sf} or \code{sfc} object which may be useful
in some limited cases.}

\item{erase}{If \code{TRUE}, apply \code{\link[tigris:erase_water]{tigris::erase_water()}} to input area and
block_xwalk before joining. Defaults to \code{FALSE}. If \code{erase} is a sf object,
the geometry of the input sf is erased from area and block_xwalk. This
option is intended to support erasing open space or other non-developed
land as well as water areas.}

\item{area_threshold}{The percentile rank cutoff of water areas to use in the erase operation, ranked by size. Defaults to 0.75, representing the water areas in the 75th percentile and up (the largest 25 percent of areas).  This value may need to be modified by the user to achieve optimal results for a given location.}

\item{keep_geometry}{If \code{TRUE}, area_xwalk is a sf object with the same
geometry as the input area. Defaults to \code{FALSE}.}

\item{crs}{Coordinate reference system to use for input data. Recommended to
set to a projected CRS if input area data is in a geographic CRS.}

\item{make_valid}{Default \code{TRUE}. If \code{TRUE}, apply \code{\link[sf:valid]{sf::st_make_valid()}} to
the input area geometry and to any sf or sfc object passed to the \code{erase}
parameter. If this has any unexpected results, set \code{make_valid = FALSE} and
prepare any invalid geometry before passing to this function.}

\item{...}{Passed to \code{\link[=make_block_xwalk]{make_block_xwalk()}}.}

\item{data}{A data frame downloaded with \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.}

\item{area_xwalk}{A area crosswalk data frame created with
\code{\link[=make_area_xwalk]{make_area_xwalk()}}. Required for \code{\link[=use_area_xwalk]{use_area_xwalk()}}.}

\item{geography}{A character string used as general description for area
geography type. Defaults to "area" but typical values could include
"neighborhood", "planning district", or "service area".}

\item{variable_col}{Variable column name. Defaults to "variable"}

\item{value_col, moe_col}{Value and margin of error column names (defaults to
"estimate" and "moe").}

\item{perc}{If \code{TRUE} (default), use the denominator column ID to calculate
each estimate as a percent share of the denominator value and use
\code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}} to calculate a new margin of error for the percent
estimate.}

\item{reliability}{If \code{TRUE}, use \code{\link[=assign_acs_reliability]{assign_acs_reliability()}} to assign a
reliability value to estimate values based on the specified \code{moe_level}.}

\item{moe_level}{The confidence level of the margin of error. Defaults to 90
(which is the same default as \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble or a sf object.
}
\description{
\code{make_area_xwalk()} creates a crosswalk data frame based on the \code{weight_col}
parameter (if \code{year = 2020}, use "POP20" for population, "HOUSING20" for
households, or "ALAND20" for land area). Using this function with other
years, requires users to add population data to the block_xwalk as the
\code{\link[tigris:blocks]{tigris::blocks()}} function only includes population and household count data
for the 2020 year. This function has also not been tested when areas include
overlapping geometry and the results may be invalid for those overlapping
areas if that is the case.
}
\details{
Using an area crosswalk

After creating an area crosswalk with \code{\link[=make_area_xwalk]{make_area_xwalk()}}, you can pass the
crosswalk to \code{\link[=use_area_xwalk]{use_area_xwalk()}} along with a data frame from
\code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} or \code{\link[=get_acs_tables]{get_acs_tables()}}. At a minimum, the data must
have a column with the same name as geoid_col along with columns named
"variable", "estimate", and "moe".

Please note that this approach to aggregation does \emph{not} work well if your
data contains "jam" values, e.g. the substitution of 0 for "1939 or older"
for the Median Year Built variable. Ideally, the weight used for aggregation
should be based on household counts when aggregating a household-level
variable and population counts when aggregating a individual-level variable.
}
\seealso{
\code{\link[tidycensus:interpolate_pw]{tidycensus::interpolate_pw()}}, \code{\link[areal:aw_interpolate]{areal::aw_interpolate()}}
}
