% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{make_block_xwalk}
\alias{make_block_xwalk}
\title{Make a crosswalk data frame for U.S. Census blocks and tracts}
\usage{
make_block_xwalk(
  state,
  county = NULL,
  year = 2020,
  by = c(TRACTCE20 = "TRACTCE"),
  keep_zipped_shapefile = TRUE,
  suffix = c("_block", "_tract"),
  ...
)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties.
Can also be a county name or vector of names.}

\item{year}{the data year; defaults to 2021}

\item{by}{A join specification created with \code{\link[dplyr:join_by]{join_by()}}, or a character
vector of variables to join by.

If \code{NULL}, the default, \verb{*_join()} will perform a natural join, using all
variables in common across \code{x} and \code{y}. A message lists the variables so
that you can check they're correct; suppress the message by supplying \code{by}
explicitly.

To join on different variables between \code{x} and \code{y}, use a \code{\link[dplyr:join_by]{join_by()}}
specification. For example, \code{join_by(a == b)} will match \code{x$a} to \code{y$b}.

To join by multiple variables, use a \code{\link[dplyr:join_by]{join_by()}} specification with
multiple expressions. For example, \code{join_by(a == b, c == d)} will match
\code{x$a} to \code{y$b} and \code{x$c} to \code{y$d}. If the column names are the same between
\code{x} and \code{y}, you can shorten this by listing only the variable names, like
\code{join_by(a, c)}.

\code{\link[dplyr:join_by]{join_by()}} can also be used to perform inequality, rolling, and overlap
joins. See the documentation at \link[dplyr:join_by]{?join_by} for details on
these types of joins.

For simple equality joins, you can alternatively specify a character vector
of variable names to join by. For example, \code{by = c("a", "b")} joins \code{x$a}
to \code{y$a} and \code{x$b} to \code{y$b}. If variable names differ between \code{x} and \code{y},
use a named character vector like \code{by = c("x_a" = "y_a", "x_b" = "y_b")}.

To perform a cross-join, generating all combinations of \code{x} and \code{y}, see
\code{\link[dplyr:cross_join]{cross_join()}}.}

\item{keep_zipped_shapefile}{Passed to \code{\link[tigris:blocks]{tigris::blocks()}} and
\code{\link[tigris:tracts]{tigris::tracts()}} to keep and re-use the zipped shapefile.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}

\item{...}{
  Arguments passed on to \code{\link[tigris:blocks]{tigris::blocks}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{make_block_xwalk()} joined U.S. Census blocks data from \code{\link[tigris:blocks]{tigris::blocks()}}
to a data frame from \code{\link[tigris:tracts]{tigris::tracts()}} to provide a crosswalk between both
geographies. If \code{year = 2020}, the suffix parameter is not used. If year is
any other year than 2020, the by parameter must be changed from the default
value of \code{c("TRACTCE20" = "TRACTCE")}. 2020 is also the only year where
\code{\link[tigris:blocks]{tigris::blocks()}} includes the population and household count data required
to use this crosswalk data frame with \code{\link[=make_area_xwalk]{make_area_xwalk()}}.
}
