% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_iteration}
\alias{race_iteration}
\title{Race or Latino Origin table codes}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{\code{code}}{Code}
\item{\code{group}}{Race or Ethnic group}
\item{\code{label}}{Short label}
}
}
\usage{
race_iteration
}
\description{
For selected tables, an alphabetic suffix follows to indicate that a table is
repeated for the nine major race and Hispanic or Latino groups.
}
\details{
\url{https://www.census.gov/programs-surveys/acs/data/data-tables/table-ids-explained.html}
}
\keyword{datasets}
