% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_jam_values.R
\name{replace_jam_values}
\alias{replace_jam_values}
\title{Replace placeholder jam values in an ACS data frame}
\usage{
replace_jam_values(data)
}
\arguments{
\item{data}{Data frame with ACS data}
}
\description{
Currently only supports variable B25035_001 from the Median Year Structure
Built table.
}
