% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_acs_cols}
\alias{select_acs_cols}
\title{Helper function to select columns typically used in an ACS table}
\usage{
select_acs_cols(
  data,
  name_col = "name",
  column_title_col = "column_title",
  est_cols = c("estimate", "moe"),
  perc_est_cols = c("perc_estimate", "perc_moe"),
  .drop_denominator = TRUE,
  ...
)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{name_col, column_title_col, est_cols, perc_est_cols}{ACS data column names
to select using \code{\link[tidyselect:all_of]{tidyselect::any_of()}}. Set any parameter to \code{NULL} to
avoid selecting columns.}

\item{.drop_denominator}{If \code{TRUE} (default), drop all columns that start
with the text "denominator".}

\item{...}{Additional parameters passed to \code{\link[dplyr:select]{dplyr::select()}}}
}
\description{
Helper function to select columns typically used in an ACS table
}
