% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_acs_variables.R
\name{summarise_acs_intensive}
\alias{summarise_acs_intensive}
\title{Summarise ACS data assuming intensive variables}
\usage{
summarise_acs_intensive(
  data,
  name_col = "NAME",
  variable_col = "variable",
  label_col = "label",
  value_col = "estimate",
  moe_col = "moe",
  na.rm = TRUE,
  digits = 2
)
}
\description{
Summarise ACS data assuming intensive variables
}
