% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tab_acs_source}
\alias{tab_acs_source}
\title{Add a source note to a table}
\usage{
tab_acs_source(
  gt_object,
  source_note = NULL,
  survey = "acs5",
  year = 2021,
  prefix = "Source: ",
  after = ".",
  tables = NULL,
  table_label = "Table",
  append_note = FALSE,
  use_md = FALSE,
  ...
)
}
\arguments{
\item{source_note}{Text to be used in the source note. We can optionally use
the \code{\link[gt:md]{md()}} and \code{\link[gt:html]{html()}} functions to style the text as Markdown or to
retain HTML elements in the text.}

\item{append_note}{If \code{TRUE}, add source_note to the end of the generated ACS
data label. If \code{FALSE}, any supplied source_note will be used instead of an
ACS label.}

\item{use_md}{If \code{TRUE}, pass source_note to gt::md() first.}
}
\description{
A wrapper for \code{\link[gt:tab_source_note]{gt::tab_source_note()}} and \code{\link[=acs_survey_label_tables]{acs_survey_label_tables()}}
}
\keyword{internal}
