% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{tab_acs_spanner_delim}
\alias{tab_acs_spanner_delim}
\title{Create new spanners using \code{gt::tab_spanner_delim()}}
\usage{
tab_acs_spanner_delim(
  data,
  column_title_col = "column_title",
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  columns = starts_with(c(value_col, moe_col, perc_prefix)),
  delim = "_",
  split = "last",
  limit = 1,
  reverse = TRUE,
  ...
)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[gt:gt]{gt()}} function.}

\item{columns}{\emph{Columns to target}

\verb{<column-targeting expression>} // \emph{default:} \code{everything()}

The columns to consider for the splitting, relabeling, and spanner setting
operations. Can either be a series of column names provided in \code{c()}, a
vector of column indices, or a select helper function (e.g.,
\code{\link[gt:starts_with]{starts_with()}}, \code{\link[gt:ends_with]{ends_with()}}, \code{\link[gt:contains]{contains()}}, \code{\link[gt:matches]{matches()}}, \code{\link[gt:num_range]{num_range()}},
and \code{\link[gt:everything]{everything()}}).}

\item{delim}{\emph{Delimiter for splitting}

\verb{scalar<character>} // \strong{required}

The delimiter text to use to split one of more column names (i.e., those
that are targeted via the \code{columns} argument).}

\item{split}{\emph{Splitting side}

\verb{singl-kw:[last|first]} // \emph{default:} \code{"last"}

Should the delimiter splitting occur from the \code{"last"} instance of the
\code{delim} character or from the \code{"first"}? The default here uses the \code{"last"}
keyword, and splitting begins at the last instance of the delimiter in the
column name. This option only has some consequence when there is a \code{limit}
value applied that is lesser than the number of delimiter characters for a
given column name (i.e., number of splits is not the maximum possible
number).}

\item{limit}{\emph{Limit for splitting}

\verb{scalar<numeric|integer|character>} // \emph{default:} \code{NULL} (\code{optional})

An optional limit to place on the splitting procedure. The default \code{NULL}
means that a column name will be split as many times are there are
delimiter characters. In other words, the default means there is no limit.
If an integer value is given to \code{limit} then splitting will cease at the
iteration given by \code{limit}. This works in tandem with \code{split} since we can
adjust the number of splits from either the right side (\code{split = "last"})
or left side (\code{split = "first"}) of the column name.}

\item{reverse}{\emph{Reverse vector of split names}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

Should the order of split names be reversed? By default, this is \code{FALSE}.}
}
\description{
Create new spanners using \code{gt::tab_spanner_delim()}
}
\keyword{internal}
