% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tigerweb_geo_index}
\alias{tigerweb_geo_index}
\title{U.S. Census Bureau ArcGIS Services Index}
\format{
A data frame with 7750 rows and 15 variables:
\describe{
\item{\code{name}}{Name}
\item{\code{type}}{Service/layer type}
\item{\code{url}}{Folder/service/layer URL}
\item{\code{urlType}}{URL type}
\item{\code{folderPath}}{Index type}
\item{\code{serviceName}}{Service name}
\item{\code{serviceType}}{Service type}
\item{\code{id}}{integer Layer ID number}
\item{\code{parentLayerId}}{integer Parent layer ID number}
\item{\code{defaultVisibility}}{logical Layer default visibility}
\item{\code{subLayerIds}}{list Sublayer ID numbers}
\item{\code{minScale}}{double Minimum scale}
\item{\code{maxScale}}{integer Maximum scale}
\item{\code{geometryType}}{Geometry type}
\item{\code{supportsDynamicLegends}}{logical Supports dynamic legends}
}
}
\usage{
tigerweb_geo_index
}
\description{
Index created with \code{\link[esri2sf:esriIndex]{esri2sf::esriIndex()}} listing all services located at
\url{https://tigerweb.geo.census.gov/arcgis/rest/services}. Access ArcGIS
services using the arcgislayers \url{https://github.com/R-ArcGIS/arcgislayers} or
esri2sf package \url{https://github.com/elipousson/esri2sf}. Last updated 2025-02-10.
}
\details{
\url{https://tigerweb.geo.census.gov/arcgis/rest/services}
}
\keyword{datasets}
