% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getACSdata.R
\name{tractACS.us}
\alias{tractACS.us}
\title{Download census-tract level ACS data for multiple states}
\usage{
tractACS.us(vars, endpoint, states)
}
\arguments{
\item{vars}{List of ACS vars to retrieve}

\item{endpoint}{ACS API endpoint}

\item{states}{List of states to download data for}
}
\value{
A data.table of selected ACS variables
}
\description{
Download census-tract level ACS data for multiple states
}
