% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_states}
\alias{usa_states}
\title{U.S. States Reference Data}
\format{
A data frame with 56 rows and 7 variables:
\describe{
\item{\code{state}}{State name}
\item{\code{state_abb}}{State USPS abbreviation}
\item{\code{STATE_GEOID}}{State GeoID}
\item{\code{division}}{Census Division name}
\item{\code{DIVISION_GEOID}}{Census Division GeoID}
\item{\code{region}}{Census Region name}
\item{\code{REGION_GEOID}}{Census Region GeoID}
}
}
\usage{
usa_states
}
\description{
A reference table of state names, abbreviations, regions, and divisions.
}
\keyword{datasets}
