% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{vec_get_acs}
\alias{vec_get_acs}
\title{Vectorized variant of \link[tidycensus:get_acs]{tidycensus::get_acs}}
\usage{
vec_get_acs(..., .fn = tidycensus::get_acs, .size = NULL, .call = caller_env())
}
\arguments{
\item{...}{Additional parameters passed to .fn.}

\item{.fn}{Function to call with parameters, Defaults to
\code{tidycensus::get_acs}. Function must require a geography parameter and
return a data frame.}

\item{.size}{Desired output size.}

\item{.call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A list of data frames (using default .fn value or another function
that returns a data frame).

A list of data frames.
}
\description{
Vectorized variant of \link[tidycensus:get_acs]{tidycensus::get_acs}
}
\examples{
\dontrun{
if (interactive()) {
  # TODO: Add examples
}
}
}
