% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_tidy.R
\name{acs_perc_cols}
\alias{acs_perc_cols}
\title{Create \code{perc} value and moe column names with prefix and separator}
\usage{
acs_perc_cols(
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = TRUE
)
}
\arguments{
\item{value_col}{Column name for estimate value column. Defaults to
"estimate".}

\item{moe_col}{Column name for margin of error column. Defaults to
"moe".}

\item{perc_prefix}{Prefix string for percent value columns.}

\item{perc_sep}{Separator string between \code{perc_prefix} and the \code{value_col}
and \code{moe_col} strings.}

\item{perc}{If \code{TRUE}, return percent value and margin of error columns.}
}
\description{
Create \code{perc} value and moe column names with prefix and separator
}
\keyword{internal}
