% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_survey.R
\name{acs_survey_match}
\alias{acs_survey_match}
\alias{acs_survey_sample}
\alias{acs_survey_ts}
\alias{acs_survey_label}
\alias{acs_survey_label_tables}
\title{Assorted helpers for ACS survey types and labels}
\usage{
acs_survey_match(survey = "acs5", error_call = caller_env())

acs_survey_sample(survey = "acs5")

acs_survey_ts(survey = "acs5", year = 2021, call = caller_env())

acs_survey_label(
  survey = "acs5",
  year = 2021,
  pattern = "{year_start}-{year} ACS {sample}-year Estimates",
  prefix = ""
)

acs_survey_label_tables(
  survey = "acs5",
  year = 2021,
  prefix = "",
  tables = NULL,
  table_label = "Table",
  sep = ", ",
  and = " and ",
  before = "",
  after = ".",
  oxford_comma = TRUE,
  ...
)
}
\arguments{
\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}
}
\description{
Assorted helpers for ACS survey types and labels
}
\keyword{internal}
