% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_ext.R
\name{acs_table_variables}
\alias{acs_table_variables}
\title{Convert a table ID into a set of variable ID using a numeric vector for
variables}
\usage{
acs_table_variables(
  table = NULL,
  variables = NULL,
  data = NULL,
  survey = "acs5",
  year = 2021,
  width = 3,
  error_call = caller_env()
)
}
\arguments{
\item{table}{A table ID string.}

\item{variables}{A numeric vector corresponding to the line number of the
variables.}

\item{data}{If data is provided and table is \code{NULL}, table is set based on
the unique values in the "table_id" column of data. If data contains more
than one table_id value, the function will error}

\item{survey}{Survey, "acs5", "acs3", or "acs1".}

\item{year}{Sample year (between 2006 and 2021).}

\item{width}{Minimum width of padded strings.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A character vector of variable ID values.
}
\description{
\code{\link[=acs_table_variables]{acs_table_variables()}} supports the creation of variable ID values based on
a table ID string. The returned variable IDs use the format returned by
\code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}, e.g. "{table_id}_{line_number}" where the
line_number is a width 3 string prefixed by "0". If variables is \code{NULL}, the
function calls \code{\link[=get_acs_metadata]{get_acs_metadata()}} with \code{metadata = "column"} and returns
all available variables for the table for the supplied year and survey.
}
