% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_table.R
\name{acs_table_variables}
\alias{acs_table_variables}
\title{Convert an ACS table ID to a set of variable ID values}
\usage{
acs_table_variables(
  table = NULL,
  variables = NULL,
  data = NULL,
  survey = "acs5",
  year = 2022,
  sep = "_",
  width = 3,
  error_call = caller_env()
)
}
\arguments{
\item{table}{An ACS table ID string.}

\item{variables}{A numeric vector corresponding to the line number of the
variables.}

\item{data}{If data is provided and table is \code{NULL}, table is set based on
the unique values in the "table_id" column of data. If data contains more
than one table_id value, the function will error}

\item{survey}{Survey, "acs5", "acs3", or "acs1".}

\item{year}{Sample year (between 2006 and 2021).}

\item{sep}{A separator character between the table ID string and variable ID
values.}

\item{width}{Variable ID suffix width.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A character vector of variable ID values for a single table.
}
\description{
\code{\link[=acs_table_variables]{acs_table_variables()}} helps to make a vector of variable ID values based on
a table ID string. The returned variable IDs use the format returned by
\code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}, e.g. "{table_id}_{line_number}" where the
line_number is a width 3 string prefixed by "0". If variables is \code{NULL}, the
function calls \code{\link[=get_acs_metadata]{get_acs_metadata()}} with \code{metadata = "column"} and returns
all available variables for the table for the supplied year and survey. Note
that the \code{sep} and \code{width} parameters should \emph{not} be changed if you are
working with data from the \code{{tidycensus}} package.
}
\examples{
acs_table_variables(table = "B15003")

acs_table_variables(table = "B15003", variables = c(1:5))
}
\seealso{
\code{\link[=acs_table_race_iteration]{acs_table_race_iteration()}}
}
