% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{assign_acs_reliability}
\alias{assign_acs_reliability}
\title{Add columns for the coefficient of variation and reliability category}
\usage{
assign_acs_reliability(
  data,
  value_col = "estimate",
  moe_col = "moe",
  moe_level = 90,
  cv_col = "cv",
  reliability_col = "reliability"
)
}
\arguments{
\item{value_col, moe_col}{Value and margin of error column names (default to
"estimate" and "moe").}

\item{moe_level}{The confidence level of the margin of error. Defaults to 90
(which is the same default as \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}).}

\item{cv_col}{Coefficient of variation column name. Defaults to "cv".}

\item{reliability_col}{Reliability category column name. Defaults to
"reliability".}
}
\description{
Add columns for the coefficient of variation and reliability category
}
