% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_acs_variables.R
\name{collapse_acs_variables}
\alias{collapse_acs_variables}
\title{Collapse variables into a new label column using \code{forcats::fct_collapse()}}
\usage{
collapse_acs_variables(
  data,
  ...,
  other_level = NULL,
  name_col = "NAME",
  variable_col = "variable",
  label_col = "label",
  value_col = "estimate",
  moe_col = "moe",
  moe_level = 90,
  reliability = FALSE,
  na.rm = TRUE,
  na_zero = TRUE,
  digits = 2,
  .add = FALSE,
  extensive = TRUE
)
}
\arguments{
\item{data}{ACS data frame input.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> A series of named character vectors. The levels in
each vector will be replaced with the name.}

\item{other_level}{Value of level used for "other" values. Always
placed at end of levels.}

\item{name_col}{Name column name, Default: 'NAME'}

\item{variable_col}{Variable column name, Default: 'variable'}

\item{label_col}{Label column name, Default: 'label'. Label is a factor
column added to the returned data frame.}

\item{value_col, moe_col}{Value and margin of error column names (default to
"estimate" and "moe").}

\item{na.rm}{Passed to \code{\link[=sum]{sum()}}, Default: \code{TRUE}}

\item{na_zero}{If \code{TRUE}, and the collapsed sum of a MOE is 0, replaced MOE
value with \code{NA}. This is beneficial for percent estimates with the margin
of error falls below 1\% and is rounded to 0 with the default number of
digits.}

\item{digits}{Passed to \code{\link[=round]{round()}}, Default: 2}

\item{.add}{When \code{FALSE}, the default, \code{group_by()} will
override existing groups. To add to the existing groups, use
\code{.add = TRUE}.

This argument was previously called \code{add}, but that prevented
creating a new grouping variable called \code{add}, and conflicts with
our naming conventions.}

\item{extensive}{Must be \code{TRUE}. If \code{FALSE} (not currently supported),
summarize collapsed variables using a weighted mean.}
}
\description{
\code{\link[=collapse_acs_variables]{collapse_acs_variables()}} uses \code{\link[forcats:fct_collapse]{forcats::fct_collapse()}} to aggregated
variables while creating a new label column. Other variables are retained in
list columns of unique values. The aggregated values for \code{perc_moe} may not
be accurate after transformation with this function. To group by additional
variables, passed a grouped data frame to data and set \code{.add = TRUE}.
}
\examples{
\dontrun{
if (interactive()) {
  edu_data <- get_acs_tables(
    "county",
    table = "B15003",
    state = "MD",
    county = "Baltimore city"
  )

  table_vars <- acs_table_variables("B15003")

  collapse_acs_variables(
    edu_data,
    "Total" = table_vars[1],
    "5th Grade or less" = table_vars[5:9],
    "6th to 8th Grade" = table_vars[10:12],
    "9th to 11th Grade" = table_vars[13:15],
    other_level = "Other"
  )
}
}
}
\seealso{
\code{\link[forcats:fct_collapse]{forcats::fct_collapse()}}, \code{\link[camiller:add_grps]{camiller::add_grps()}}
}
