% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_acs.R
\name{fmt_acs_county}
\alias{fmt_acs_county}
\alias{fmt_acs_minutes}
\title{Format place names or column titles in a gt table or data frame with ACS data}
\usage{
fmt_acs_county(
  data,
  state = NULL,
  pattern = ", {state}",
  replacement = "",
  name_col = "NAME",
  columns = all_of(name_col),
  ...
)

fmt_acs_minutes(
  data,
  pattern = "[:space:]minutes$",
  replacement = "",
  column_title_col = "column_title",
  columns = all_of(column_title_col),
  ...
)
}
\arguments{
\item{data}{\emph{The gt table data object}

\verb{obj:<gt_tbl>} // \strong{required}

This is the \strong{gt} table object that is commonly created through use of the
\code{\link[gt:gt]{gt()}} function.}

\item{state}{State name. Required if state is included in pattern.}

\item{pattern}{Passed to \code{\link[glue:glue]{glue::glue()}} and \code{\link[stringr:str_replace]{stringr::str_replace()}} for
\code{\link[=fmt_acs_county]{fmt_acs_county()}} or just to \code{\link[stringr:str_replace]{stringr::str_replace()}} by
\code{\link[=fmt_acs_minutes]{fmt_acs_minutes()}}. Defaults to \code{", {state}"} which strips the state name
from a column of county-level name values or \code{"[:space:]minutes$"} which
strips the trailing text for minutes.}

\item{replacement}{Passed to \code{\link[stringr:str_replace]{stringr::str_replace()}}. Defaults to \code{""}.}

\item{name_col}{Name for column with place name values. Defaults to "NAME"}

\item{columns}{\emph{Columns to target}

\code{\link[gt:rows-columns]{<column-targeting expression>}} // \emph{default:} \code{everything()}

Can either be a series of column names provided in \code{c()}, a vector of
column indices, or a select helper function (e.g. \code{\link[gt:starts_with]{starts_with()}},
\code{\link[gt:ends_with]{ends_with()}}, \code{\link[gt:contains]{contains()}}, \code{\link[gt:matches]{matches()}}, \code{\link[gt:num_range]{num_range()}} and \code{\link[gt:everything]{everything()}}).}

\item{...}{
  Arguments passed on to \code{\link[gt:fmt]{gt::fmt}}
  \describe{
    \item{\code{rows}}{\emph{Rows to target}

\code{\link[gt:rows-columns]{<row-targeting expression>}} // \emph{default:} \code{everything()}

In conjunction with \code{columns}, we can specify which of their rows should
undergo formatting. The default \code{\link[gt:everything]{everything()}} results in all rows in
\code{columns} being formatted. Alternatively, we can supply a vector of row
captions within \code{c()}, a vector of row indices, or a select helper
function (e.g. \code{\link[gt:starts_with]{starts_with()}}, \code{\link[gt:ends_with]{ends_with()}}, \code{\link[gt:contains]{contains()}}, \code{\link[gt:matches]{matches()}},
\code{\link[gt:num_range]{num_range()}}, and \code{\link[gt:everything]{everything()}}). We can also use expressions to filter
down to the rows we need (e.g., \verb{[colname_1] > 100 & [colname_2] < 50}).}
    \item{\code{compat}}{\emph{Formatting compatibility}

\verb{vector<character>} // \emph{default:} \code{NULL} (\code{optional})

An optional vector that provides the compatible classes for the formatting.
By default this is \code{NULL}.}
    \item{\code{fns}}{\emph{Formatting functions}

\verb{function|list of functions} // \strong{required}

Either a single formatting function or a named list of functions. Can also
be anonymous functions, in both base R (\verb{\\(x) x + 1}) and \code{rlang}
(\code{~.x + 1}) syntax.}
  }}

\item{column_title_col}{Column title column.}
}
\description{
\code{\link[=fmt_acs_county]{fmt_acs_county()}} is helpful for stripping the state name from county-level
ACS data and \code{\link[=fmt_acs_minutes]{fmt_acs_minutes()}} does the same for a column with a duration
(e.g. commute times). If data is not a \code{gt_tbl} object, both function can use
\code{\link[dplyr:mutate]{dplyr::mutate()}} to transform a standard data frame.
}
\keyword{gt}
