% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_acs}
\alias{geom_acs}
\alias{aes_errorbarh}
\alias{aes_errorbarv}
\alias{geom_acs_errorbar}
\alias{geom_acs_errorbarh}
\alias{geom_acs_errorbarv}
\title{\code{geom_} functions for plotting ACS data with ggplot2}
\usage{
aes_errorbarh(
  xmin = NULL,
  xmax = NULL,
  ...,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = FALSE
)

aes_errorbarv(
  ymin = NULL,
  ymax = NULL,
  ...,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = FALSE
)

geom_acs_errorbar(
  mapping = NULL,
  data = NULL,
  ...,
  na.rm = TRUE,
  orientation = NA,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = FALSE
)

geom_acs_errorbarh(
  mapping = NULL,
  data = NULL,
  ...,
  na.rm = TRUE,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = FALSE
)

geom_acs_errorbarv(
  mapping = NULL,
  data = NULL,
  ...,
  na.rm = TRUE,
  value_col = "estimate",
  moe_col = "moe",
  perc_prefix = "perc",
  perc_sep = "_",
  perc = FALSE
)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
These functions wrap existing \code{{ggplot2}} functions and provide default
aesthetic mappings that work with the structure of input ACS data to support
shorter, less repetitive plotting code.
}
\keyword{ggplot2}
\keyword{internal}
