% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus.R
\name{get_acs_tables}
\alias{get_acs_tables}
\title{Get multiple ACS tables}
\usage{
get_acs_tables(
  geography,
  tables = NULL,
  cache_table = TRUE,
  year = 2021,
  survey = "acs5",
  label = FALSE,
  perc = TRUE,
  geoid = "GEOID",
  ...
)
}
\arguments{
\item{geography}{The geography of your data.}

\item{tables}{A character vector of tables.}

\item{cache_table}{Whether or not to cache table names for faster future
access. Defaults to FALSE; if TRUE, only needs to be called once per
dataset.  If variables dataset is already cached via the
\code{load_variables} function, this can be bypassed.}

\item{year}{The year, or endyear, of the ACS sample. 5-year ACS data is
available from 2009 through 2021; 1-year ACS data is available from 2005
through 2021, with the exception of 2020.  Defaults to 2021.}

\item{survey}{The ACS contains one-year, three-year, and five-year surveys
expressed as "acs1", "acs3", and "acs5". The default selection is "acs5."}

\item{...}{Other keyword arguments}
}
\description{
Get multiple ACS tables
}
\keyword{internal}
