% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_survey_ts.R
\name{get_acs_ts}
\alias{get_acs_ts}
\alias{acs_survey_ts}
\title{Get a time series of ACS data}
\usage{
get_acs_ts(
  geography,
  state = NULL,
  survey = "acs5",
  year = 2021,
  cache_table = TRUE,
  ...
)

acs_survey_ts(survey, year)
}
\description{
Get a time series of ACS data
}
