% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt_acs}
\alias{gt_acs}
\title{Format American Community Survey estimate and percent estimate columns for a
gt table}
\usage{
gt_acs(
  data,
  ...,
  est_cols = c("estimate", "moe"),
  est_col_label = "Est.",
  perc_cols = c("perc_estimate", "perc_moe"),
  perc_col_label = "\% share",
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  decimals = 0,
  column_title_col = "column_title",
  column_title_label = NULL,
  name_col = "NAME",
  name_col_label = NULL,
  source_note = NULL,
  append_note = FALSE,
  survey = "acs5",
  year = 2021,
  table = NULL,
  prefix = "Source: ",
  end = "."
)
}
\arguments{
\item{data}{\emph{Input data table}

\verb{obj:<data.frame>|obj:<tbl_df>} // \strong{required}

A \code{data.frame} object or a tibble (\code{tbl_df}).}

\item{...}{Additional parameters passed to \code{\link[gt:fmt_number]{gt::fmt_number()}} by
\code{\link[=fmt_acs_estimate]{fmt_acs_estimate()}} or to \code{\link[gt:fmt_percent]{gt::fmt_percent()}} by \code{\link[=fmt_acs_percent]{fmt_acs_percent()}}.}

\item{est_cols, est_col_label, est_spanner}{Passed to columns, col_labels, and
spanner parameter for \code{\link[=fmt_acs_estimate]{fmt_acs_estimate()}}}

\item{perc_cols, perc_col_label, perc_spanner}{Passed to columns, col_labels,
and spanner parameter for \code{\link[=fmt_acs_percent]{fmt_acs_percent()}}}

\item{combined_spanner}{If not \code{NULL}, combined_spanner is passed to label
parameter of \code{\link[gt:tab_spanner]{gt::tab_spanner()}} using both est_cols and perc_cols as the
columns parameter.}

\item{decimals}{\emph{Number of decimal places}

\code{scalar<numeric|integer>(val>=0)} // \emph{default:} \code{2}

This corresponds to the exact number of decimal places to use. A value
such as \code{2.34} can, for example, be formatted with \code{0} decimal places and
it would result in \code{"2"}. With \code{4} decimal places, the formatted value
becomes \code{"2.3400"}. The trailing zeros can be removed with
\code{drop_trailing_zeros = TRUE}. If you always need \code{decimals = 0}, the
\code{\link[gt:fmt_integer]{fmt_integer()}} function should be considered.}

\item{column_title_col}{Column name to check for in the input gt table. Only
required if column_title_label is supplied.}

\item{column_title_label}{Optional label applied to column named in
column_title_col.}

\item{source_note}{\emph{Source note text}

\verb{scalar<character>} // \strong{required}

Text to be used in the source note. We can optionally use the \code{\link[gt:md]{md()}} and
\code{\link[gt:html]{html()}} functions to style the text as Markdown or to retain HTML elements
in the text.}

\item{append_note}{If \code{TRUE}, add source_note to the end of the generated ACS
data label. If \code{FALSE}, any supplied source_note will be used instead of an
ACS label.}

\item{survey}{ACS survey, "acs5", "acs3", or "acs1".}

\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}

\item{table}{One or more table IDs to include in label or source note.}

\item{prefix}{Text to insert before ACS survey label.}
}
\description{
Create or format a gt table with an estimate and margin of error and
(optionally) percent estimate and margin of error value. Use in combination
with the \code{\link[=select_acs_cols]{select_acs_cols()}} helper function to prep data before creating a
table.
}
\examples{
\dontrun{
if (interactive()) {
  data <- get_acs_tables(
    geography = "county",
    county = "Baltimore city",
    state = "MD",
    table = "B08134"
  )

  tbl_data <- filter_acs(data, indent == 1, line_number <= 10)
  tbl_data <- select_acs_cols(tbl_data)

  gt_acs(
    tbl_data,
    column_title_label = "Commute time",
    table = "B08134"
  )
}
}
}
\seealso{
Other gt table: 
\code{\link{cols_merge_uncert_ext}()},
\code{\link{fmt_acs_estimate}()},
\code{\link{tab_acs_source_note}()}
}
\concept{gt table}
