% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt.R
\name{gt_census_cols}
\alias{gt_census_cols}
\title{Format Census estimate and percent estimate columns}
\usage{
gt_census_cols(
  gt_object,
  est_cols = c("estimate", "moe"),
  est_col_label = "Est.",
  perc_cols = c("perc_estimate", "perc_moe"),
  perc_col_label = "\% share",
  est_spanner = NULL,
  perc_spanner = NULL,
  combined_spanner = NULL,
  column_title = NULL,
  source = "Source: 2017-2021 ACS 5-year Estimates",
  tables = NULL,
  decimals = 0
)
}
\description{
Format Census estimate and percent estimate columns
}
\keyword{internal}
