% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_acs.R
\name{join_acs_denominator}
\alias{join_acs_denominator}
\title{Join denominator values based on a supplied denominator column}
\usage{
join_acs_denominator(
  data,
  geoid_col = "GEOID",
  value_col = "estimate",
  moe_col = "moe",
  column_id_col = "column_id",
  column_title_col = "column_title",
  denominator_col = NULL,
  denominator_prefix = "denominator_",
  na_matches = "never",
  digits = 2,
  call = caller_env()
)
}
\arguments{
\item{data}{A data frame with column names including "column_id",
"column_title", "denominator_column_id", "estimate", and "moe".}

\item{geoid_col}{A GeoID column name to use if perc is \code{TRUE}, Defaults to
'GEOID'.}

\item{value_col}{Value column name}

\item{moe_col}{Margin of error column name}

\item{column_id_col}{Column ID column name from Census Reporter metadata.
Defaults to "column_id"}

\item{denominator_col}{Denominator column ID name from Census Reporter
metadata. Defaults to \code{NULL}}

\item{denominator_prefix}{Prefix to use for denominator column names.}
}
\description{
Note that this function and the related \code{\link[=join_acs_percent]{join_acs_percent()}} function depends
on the column-level metadata supplied by \code{\link[=label_acs_metadata]{label_acs_metadata()}}.
}
