% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_acs.R
\name{join_acs_percent}
\alias{join_acs_percent}
\title{Join percent estimates to ACS data based on denominator values}
\usage{
join_acs_percent(
  data,
  geoid_col = "GEOID",
  column_id_col = "column_id",
  denominator_col = NULL,
  denominator_prefix = "denominator_",
  value_col = "estimate",
  moe_col = "moe",
  na_matches = "never",
  digits = 2
)
}
\arguments{
\item{data}{A data frame with column names including "column_id",
"column_title", "denominator_column_id", "estimate", and "moe".}

\item{geoid_col}{A GeoID column name to use if perc is \code{TRUE}, Defaults to
'GEOID'.}

\item{column_id_col}{Column ID column name from Census Reporter metadata.
Defaults to "column_id"}

\item{denominator_col}{Denominator column ID name from Census Reporter
metadata. Defaults to \code{NULL}}

\item{denominator_prefix}{Prefix to use for denominator column names.}

\item{value_col}{Value column name}

\item{moe_col}{Margin of error column name}

\item{na_matches}{Should two \code{NA} or two \code{NaN} values match?
\itemize{
\item \code{"na"}, the default, treats two \code{NA} or two \code{NaN} values as equal, like
\code{\%in\%}, \code{\link[=match]{match()}}, and \code{\link[=merge]{merge()}}.
\item \code{"never"} treats two \code{NA} or two \code{NaN} values as different, and will
never match them together or to any other values. This is similar to joins
for database sources and to \code{base::merge(incomparables = NA)}.
}}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}
}
\description{
\code{\link[=join_acs_percent]{join_acs_percent()}} uses the denominator_column_id value from the column
metadata added with \code{\link[=label_acs_metadata]{label_acs_metadata()}} to calculate the estimate as a
percent share of the denominator value. \code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}} is used to
calculate the margin of error for the percentage.
}
\seealso{
\code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}}, \code{\link[camiller:calc_shares]{camiller::calc_shares()}}
}
