% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs.R
\name{join_acs_percent}
\alias{join_acs_percent}
\title{Calculate estimates as a percent of the denominator estimates}
\usage{
join_acs_percent(data, geoid = "GEOID", na_matches = "never", digits = 2)
}
\arguments{
\item{na_matches}{Should two \code{NA} or two \code{NaN} values match?
\itemize{
\item \code{"na"}, the default, treats two \code{NA} or two \code{NaN} values as equal, like
\code{\%in\%}, \code{\link[=match]{match()}}, and \code{\link[=merge]{merge()}}.
\item \code{"never"} treats two \code{NA} or two \code{NaN} values as different, and will
never match them together or to any other values. This is similar to joins
for database sources and to \code{base::merge(incomparables = NA)}.
}}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}
}
\description{
Use the denominator_column_id value from the column metadata added with
\code{\link[=label_acs_metadata]{label_acs_metadata()}} to calculate the estimate as a percent share of the
denominator value. \code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}} is used to calculate the margin
of error for the percentage.
}
\seealso{
\code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}}
}
\keyword{internal}
