% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_acs_metadata.R
\name{label_acs_metadata}
\alias{label_acs_metadata}
\alias{label_acs_table_metadata}
\alias{label_acs_column_metadata}
\title{Label American Community Survey data using table and column metadata from
Census Reporter}
\usage{
label_acs_metadata(
  data,
  survey = "acs5",
  year = 2022,
  perc = TRUE,
  geoid_col = "GEOID",
  variable_col = "variable"
)

label_acs_table_metadata(
  data,
  survey = "acs5",
  year = 2022,
  variable_col = "variable",
  table_id_col = "table_id",
  call = caller_env()
)

label_acs_column_metadata(
  data,
  survey = "acs5",
  year = 2022,
  variable_col = "variable",
  column_id_col = "column_id",
  column_title_col = "column_title",
  table_id_col = "table_id",
  call = caller_env()
)
}
\arguments{
\item{data}{A data frame downloaded with \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.}

\item{survey}{Survey, "acs5", "acs3", or "acs1".}

\item{year}{Sample year (between 2006 and 2022).}

\item{perc}{If \code{TRUE} (default), use the denominator column ID to calculate
each estimate as a percent share of the denominator value and use
\code{\link[tidycensus:moe_prop]{tidycensus::moe_prop()}} to calculate a new margin of error for the percent
estimate.}

\item{geoid_col}{A GeoID column name to use if perc is \code{TRUE}, Defaults to
'GEOID'.}

\item{variable_col}{Variable column name. Defaults to "variable"}

\item{column_id_col}{Column ID column name from Census Reporter metadata.
Defaults to "column_id"}
}
\description{
Label American Community Survey data using table and column metadata from
Census Reporter
}
\seealso{
\code{\link[=join_acs_percent]{join_acs_percent()}}
}
\keyword{internal}
