% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{labs_acs_survey}
\alias{labs_acs_survey}
\title{Label a ggplot2 plot and add a caption based on an ACS survey year}
\usage{
labs_acs_survey(
  ...,
  caption = NULL,
  survey = "acs5",
  year = 2021,
  prefix = "Source: ",
  table = NULL
)
}
\arguments{
\item{survey}{ACS survey, "acs5", "acs3", or "acs1".}

\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}

\item{prefix}{Text to insert before ACS survey label.}

\item{table}{One or more table IDs to include in label or source note.}
}
\description{
Label a ggplot2 plot and add a caption based on an ACS survey year
}
\keyword{internal}
