% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{labs_acs_survey}
\alias{labs_acs_survey}
\title{Label a ggplot2 plot and add a caption based on an ACS survey year}
\usage{
labs_acs_survey(
  ...,
  caption = NULL,
  survey = "acs5",
  year = 2021,
  prefix = "Source: ",
  tables = NULL
)
}
\arguments{
\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}
}
\description{
Label a ggplot2 plot and add a caption based on an ACS survey year
}
\keyword{internal}
