% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{labs_acs_survey}
\alias{labs_acs_survey}
\title{Label a ggplot2 plot and add a caption based on an ACS survey year}
\usage{
labs_acs_survey(
  ...,
  caption = NULL,
  survey = "acs5",
  year = 2022,
  prefix = "Source: ",
  table = NULL
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:labs]{ggplot2::labs}}
  \describe{
    \item{\code{title}}{The text for the title.}
    \item{\code{subtitle}}{The text for the subtitle for the plot which will be
displayed below the title.}
    \item{\code{tag}}{The text for the tag label which will be displayed at the
top-left of the plot by default.}
    \item{\code{alt,alt_insight}}{Text used for the generation of alt-text for the plot.
See \link[ggplot2]{get_alt_text} for examples.}
  }}

\item{caption}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}

\item{survey}{ACS survey, "acs5", "acs3", or "acs1".}

\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}

\item{prefix}{Text to insert before ACS survey label.}

\item{table}{One or more table IDs to include in label or source note.}
}
\description{
\code{labs_acs_survey()} uses \code{acs_survey_label_table()} to create a label for a
ggplot2 plot passed to the caption parameter of \code{\link[ggplot2:labs]{ggplot2::labs()}}.
}
\keyword{ggplot2}
