% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{make_block_xwalk}
\alias{make_block_xwalk}
\title{Make crosswalk data for U.S. Census blocks and tracts}
\usage{
make_block_xwalk(
  state,
  county = NULL,
  year = 2020,
  area = NULL,
  type = c("block-to-tract", "block-to-area-join", "block-to-area-intersection"),
  by = c(TRACTCE20 = "TRACTCE"),
  suffix = c("_block", "_tract"),
  crs = NULL,
  join = sf::st_intersects,
  left = TRUE,
  largest = FALSE,
  keep_zipped_shapefile = TRUE,
  ...
)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties.
Can also be a county name or vector of names.}

\item{year}{the data year; defaults to 2024}

\item{area}{A sf object. Required if \code{type} is  "block-to-area-join or
"block-to-area-intersection".}

\item{type}{One of three types of \code{"block-to-tract"}, \code{"block-to-area-join"}, or \code{"block-to-area-intersection"}}

\item{by}{Specification of join variables in the format of c("block column
name for tract" = "tract column name"). Passed to \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}.}

\item{suffix}{Suffixes added to the output to disambiguate column names from
the block and tract data. Unused for 2020 data.}

\item{crs}{Coordinate reference system to return.}

\item{join}{geometry predicate function with the same profile as \link[sf]{st_intersects}; see details}

\item{left}{logical; if \code{TRUE} return the left join, otherwise an inner join; see details.
see also \link[dplyr:mutate-joins]{left_join}}

\item{largest}{logical; if \code{TRUE}, return \code{x} features augmented with the fields of \code{y} that have the largest overlap with each of the features of \code{x}; see https://github.com/r-spatial/sf/issues/578}

\item{keep_zipped_shapefile}{Passed to \code{\link[tigris:blocks]{tigris::blocks()}} and
\code{\link[tigris:tracts]{tigris::tracts()}} to keep and re-use the zipped shapefile.}

\item{...}{
  Arguments passed on to \code{\link[tigris:blocks]{tigris::blocks}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{\link[=make_block_xwalk]{make_block_xwalk()}} joined U.S. Census blocks data from \code{\link[tigris:blocks]{tigris::blocks()}}
to a data frame from \code{\link[tigris:tracts]{tigris::tracts()}} to provide a crosswalk between both
geographies. If \code{year = 2020}, the suffix parameter is not used. If year is
any other year than 2020, the by parameter must be changed from the default
value of \code{c("TRACTCE20" = "TRACTCE")}. 2020 is also the only year where
\code{\link[tigris:blocks]{tigris::blocks()}} includes the population and household count data required
to use this crosswalk data frame with \code{\link[=make_area_xwalk]{make_area_xwalk()}}.
}
