% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk.R
\name{make_block_xwalk}
\alias{make_block_xwalk}
\title{Make crosswalk data for U.S. Census blocks and tracts}
\usage{
make_block_xwalk(
  state,
  county = NULL,
  year = 2020,
  by = c(TRACTCE20 = "TRACTCE"),
  keep_zipped_shapefile = TRUE,
  suffix = c("_block", "_tract"),
  crs = NULL,
  ...
)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties. Can
also be a county name or vector of names.}

\item{year}{the data year; defaults to 2022}

\item{by}{Specification of join variables in the format of c("block column
name for tract" = "tract column name"). Passed to \code{\link[dplyr:mutate-joins]{dplyr::left_join()}}.}

\item{keep_zipped_shapefile}{Passed to \code{\link[tigris:blocks]{tigris::blocks()}} and
\code{\link[tigris:tracts]{tigris::tracts()}} to keep and re-use the zipped shapefile.}

\item{suffix}{Suffixes added to the output to disambiguate column names from
the block and tract data. Unused for 2020 data.}

\item{crs}{Coordinate reference system to return.}

\item{...}{
  Arguments passed on to \code{\link[tigris:blocks]{tigris::blocks}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\code{\link[=make_block_xwalk]{make_block_xwalk()}} joined U.S. Census blocks data from \code{\link[tigris:blocks]{tigris::blocks()}}
to a data frame from \code{\link[tigris:tracts]{tigris::tracts()}} to provide a crosswalk between both
geographies. If \code{year = 2020}, the suffix parameter is not used. If year is
any other year than 2020, the by parameter must be changed from the default
value of \code{c("TRACTCE20" = "TRACTCE")}. 2020 is also the only year where
\code{\link[tigris:blocks]{tigris::blocks()}} includes the population and household count data required
to use this crosswalk data frame with \code{\link[=make_area_xwalk]{make_area_xwalk()}}.
}
