% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_acs}
\alias{scale_acs}
\alias{scale_x_acs}
\alias{scale_y_acs}
\alias{scale_x_acs_estimate}
\alias{scale_y_acs_percent}
\alias{scale_x_acs_percent}
\alias{scale_y_acs_estimate}
\alias{scale_x_acs_ts}
\alias{scale_y_acs_ts}
\title{Scales for plotting ACS data with ggplot2}
\usage{
scale_x_acs(..., perc = FALSE)

scale_y_acs(..., perc = FALSE)

scale_x_acs_estimate(name = "Estimate", ..., labels = scales::label_comma())

scale_y_acs_percent(
  name = "Est. \% of total",
  ...,
  labels = scales::label_percent()
)

scale_x_acs_percent(
  name = "Est. \% of total",
  ...,
  labels = scales::label_percent()
)

scale_y_acs_estimate(name = "Estimate", ..., labels = scales::label_comma())

scale_x_acs_ts(name = "Year", ..., breaks = NULL, survey = "acs5", year = 2022)

scale_y_acs_ts(name = "Year", ..., breaks = NULL, survey = "acs5", year = 2022)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}

\item{perc}{If \code{TRUE}, use the \code{scale_x_acs_percent} or
\code{scale_y_acs_percent}. Defaults to \code{FALSE}.}

\item{name}{The name of the scale. Used as the axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}

\item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item An expression vector (must be the same length as breaks). See ?plotmath for details.
\item A function that takes the breaks as input and returns labels
as output. Also accepts rlang \link[rlang:as_function]{lambda} function
notation.
}}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:new_transform]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Note that for position scales, limits are provided after scale expansion.
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{survey}{ACS survey, "acs5", "acs3", or "acs1".}

\item{year}{Based on the year and survey, \code{\link[=acs_survey_ts]{acs_survey_ts()}} returns a vector
of years for non-overlapping ACS samples to allow comparison.}
}
\description{
Scales for plotting ACS data with ggplot2
}
\keyword{ggplot2}
